// backend/src/modules/shared/services/config/config.enum.ts
export enum Config {
  DB_CONNECTION = 'DB_CONNECTION',
  DB_DATABASE = 'DB_DATABASE',
  DB_HOST = 'DB_HOST',
  DB_SCHEMA = 'DB_SCHEMA',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_PORT = 'DB_PORT',

  SERVER_URL = 'SERVER_URL',

  PUBLIC_CRYPTO_ALGORITHM = 'PUBLIC_CRYPTO_ALGORITHM',
  PUBLIC_CRYPTO_KEY = 'PUBLIC_CRYPTO_KEY',

  API_KEY = 'API_KEY',
  JWT_ALGORITHM = 'JWT_ALGORITHM',
  JWT_ID_TOKEN_SECRET = 'JWT_ID_TOKEN_SECRET',
  JWT_ID_TOKEN_EXP_TIME = 'JWT_ID_TOKEN_EXP_TIME',
  JWT_REFRESH_TOKEN_SECRET = 'JWT_REFRESH_TOKEN_SECRET',
  JWT_REFRESH_TOKEN_EXP_TIME = 'JWT_REFRESH_TOKEN_EXP_TIME',

  GOOGLE_CLIENT_ID = 'GOOGLE_CLIENT_ID',
  GOOGLE_CLIENT_SECRET = 'GOOGLE_CLIENT_SECRET',
  GOOGLE_REDIRECT_URL = 'GOOGLE_REDIRECT_URL',
}